Step III: Stepwise on the consolidated output from different GDS runs

Runs the stepwise regression on the output received from top models of the consolidated output of different GDS runs. With n being the number of runs, the stepwise regression starts with at most (n-3) selected effects from the previous step. The remaining effects from the previous step as well as all main effects are given a chance to enter into the model using the forward-backward stepwise regression.

StepIII_stepwise(
  xstart,
  xremain,
  Xmain,
  Xint,
  Y,
  cri.penter = 0.01,
  cri.premove = 0.05,
  opt.heredity = "none"
)

Source

Singh, R. and Stufken, J. (2022). Factor selection in screening experiments by aggregation over random models, 1--31. https://arxiv.org/abs/2205.13497/

Arguments

xstart

a vector with effects' names corresponding to the starting model.

xremain

a vector with effects' names corresponding to the remaining main effects and other effects that needs to be explored with stepwise regression.

Xmain

a \(n \times m\) matrix of m main effects.

Xint

a matrix of m choose 2 two-factor interactions.

Y

a vector of n responses.

cri.penter

the p-value cutoff for an effect to enter into the stepwise regression model

cri.premove

the p-value cutoff for an effect to exit from the stepwise regression model

opt.heredity

a string with either none, or weak, or strong. Denotes whether the effect-heredity (weak or strong) should be embedded in GDS-ARM. The default value is none as suggested in Singh and Stufken (2022).

Value

A list returning the effects identified as active as well as the corresponding identified important factors.