Step I: Multiple GDS runs with random interactions

Runs Gauss Dantzig Selector (GDS) multiple times, each time with a differently set of randomly selected two-factor interactions. All m main effects are included in each GDS run. For each set of randomly selected interactions, the best GDS output is chosen among delta.n values of delta. We use kmeans with 2 clusters and BIC to select such best model.

StepI_chooseints(
  delta.n = 10,
  nint,
  nrep,
  Xmain,
  Xint,
  Y,
  opt.heredity = c("none")
)

Source

Singh, R. and Stufken, J. (2022). Factor selection in screening experiments by aggregation over random models, 1--31. https://arxiv.org/abs/2205.13497/

Arguments

delta.n

a positive integer suggesting the number of delta values to be tried. delta.n values of delta will be used strictly between 0 and max(t(X)y). The default value is set to 10.

nint

a positive integer representing the number of randomly chosen interactions. The suggested value to use is the ceiling of 20% of the total number of interactions, that is, for m factors, we have ceiling(0.2(m choose 2)).

nrep

a positive integer representing the number of times GDS should be run. The suggested value is (m choose 2).

Xmain

a \(n \times m\) matrix of m main effects.

Xint

a matrix of \({m \choose 2}\) two-factor interactions.

Y

a vector of n responses.

opt.heredity

a string with either none, or weak, or strong. Denotes whether the effect-heredity (weak or strong) should be embedded in GDS-ARM. The default value is none as suggested in Singh and Stufken (2022).

Value

A list containing the (a) matrix of the output of each GDS run with each row representing the selected effects from the corresponding GDS run, (b) a vector with the corresponding BIC values of each model.